Todo sirve sin manager

+++ src/CasoII/ICobro.java
@@ -2,7 +2,7 @@ package CasoII;
 
 public interface ICobro {
 
-    public double cobro(Cliente c);
+    public float cobro(Cliente c);
 
     public void total(Cliente c);
